Cache the background.
The math to draw the ticks takes a surprisingly long time. I've cached the background (including the ticks so that they're redrawn less often.

This was the lowest-hanging optimisation fruit. This cut the rendering time roughly in half. The next-lowest-hanging fruit is reading the preferences only when they change instead of on every frame.

issue #35